handle nested segwit address requests in proxy
