+++ src/comps/temp-withdraw.tsx
@@ -24,25 +24,15 @@ import {
   callContractXverse,
 } from "@/util/wallet-utils";
 import { getStacksNetwork } from "@/util/get-stacks-network";
-import { validate, Network, getAddressInfo } from "bitcoin-address-validation";
+
 import { decodeBitcoinAddress } from "@/util/decode-bitcoin-address";
+import { validateBitcoinAddress } from "@/util/validate-bitcoin-address";
 
 const decodeBitcoinAddressToClarityRecipient = (
   address: string,
   network: DefaultNetworkConfigurations,
 ) => {
-  const isMainnet = network === "mainnet";
-  const addressInfo = getAddressInfo(address);
-  let addressNetwork = isMainnet ? Network.mainnet : Network.regtest;
-  // bech32 is different between all three networks
-  // base58 is same between regtest and testnet
-  // the validation lib defaults to testnet when using base58 addresses the same address is valid for regtest
-  // so it would break if the passed address is regtest base58 and the network is regtest,
-  // since to it the address is testnet by default, this is basically a workaround for regtest
-  const isRegtestBase58 = !addressInfo.bech32 && !isMainnet;
-  addressNetwork = isRegtestBase58 ? Network.testnet : addressNetwork;
-
-  if (validate(address, addressNetwork)) {
+  if (validateBitcoinAddress(address, network)) {
     return decodeBitcoinAddress(address);
   } else {
     throw new Error(`Invalid address: ${address}`);
