+++ src/env.ts
@@ -16,9 +16,8 @@ export const env = {
 
   BITCOIN_RPC_USER_NAME: process.env.BITCOIN_RPC_USER_NAME || "devnet",
   BITCOIN_RPC_PASSWORD: process.env.BITCOIN_RPC_PASSWORD || "devnet",
-  WALLET_NETWORK: (process.env.WALLET_NETWORK || "sbtcDevenv") as
-    | DefaultNetworkConfigurations
-    | undefined,
+  WALLET_NETWORK: (process.env.WALLET_NETWORK ||
+    "sbtcDevenv") as DefaultNetworkConfigurations,
   SBTC_CONTRACT_DEPLOYER: process.env.SBTC_CONTRACT_DEPLOYER,
   BANNER_CONTENT: process.env.BANNER_CONTENT,
   RECLAIM_LOCK_TIME: process.env.RECLAIM_LOCK_TIME,
